Add year to actions

+++ src/app/lib/actions.ts
@@ -1,40 +1,27 @@
 'use server';
 
 import { revalidatePath } from 'next/cache';
-import { deleteResultByKey, getResultItems, insertResult, isAdmin, updateAvatar } from '../db/data';
+import { deleteResultByKey, getResultItems, isAdmin } from '../db/data';
 import { auth } from '@/app/lib/auth';
 import { isNullOrEmpty } from '@/app/utils';
-import { type Result, type ResultItem, type ResultType } from '@/app/lib/types';
+import { type ResultItem } from '@/app/lib/types';
 import { env } from '@/app/lib/env';
 
-export async function deleteResult(key: string) {
+export async function deleteResult(key: string, year: number) {
+  if (!(await canEditResults({ year }))) {
+    return {
+      error: 'Hiba történt a törlés során! Nincs jogosultság a törléshez.',
+    };
+  }
   await deleteResultByKey(key);
-  revalidatePath('/eredmenyek');
+  revalidatePath(`/${year}/eredmenyek`);
+  return null;
 }
 
-export async function uploadResult(result: { key: string; result: Result; type: ResultType }) {
-  await insertResult({ ...result, year: 0 });
-  revalidatePath('/eredmenyek');
-}
-
-export async function canEditResults() {
+export async function canEditResults({ year }: Readonly<{ year: number }>) {
   const session = await auth();
   const email = session?.user?.email;
-  return !isNullOrEmpty(email) && (await isAdmin(email));
-}
-
-export async function updateUserAvatar({
-  id,
-  avatar,
-}: Readonly<{
-  id: string | null | undefined;
-  avatar: string | null;
-  provider: string | null | undefined;
-}>) {
-  if (isNullOrEmpty(id)) {
-    return;
-  }
-  return await updateAvatar({ id, avatar: avatar ?? null });
+  return !isNullOrEmpty(email) && (await isAdmin(email, year));
 }
 
 export async function getResults(sportag: string): Promise<ResultItem[]> {
